Code cleanup, test corrections, and linting.
Files checked in: CONTRIBUTING.md DataRepo/advanced_search_utils.py DataRepo/tasks.py DataRepo/templates/DataRepo/search/downloads/download_progress.html DataRepo/tests/test_views.py DataRepo/urls.py DataRepo/views.py TraceBase/__init__.py TraceBase/celery.py TraceBase/settings.py DataRepo/tests/test_advanced_search_utils.py
