+++ DataRepo/tests/test_views.py
@@ -22,19 +22,7 @@ from DataRepo.models import (
     get_all_models,
 )
 from DataRepo.tests.tracebase_test_case import TracebaseTestCase
-from DataRepo.views import (
-    DataValidationView,
-    constructAdvancedQuery,
-    createNewBasicQuery,
-    getAllBrowseData,
-    getJoinedRecFieldValue,
-    isQryObjValid,
-    isValidQryObjPopulated,
-    pathStepToPosGroupType,
-    performQuery,
-    rootToFormatInfo,
-    searchFieldToDisplayField,
-)
+from DataRepo.views import DataValidationView
 
 
 class ViewTests(TracebaseTestCase):
@@ -535,223 +523,6 @@ class ViewTests(TracebaseTestCase):
         self.assertTemplateUsed(response, "DataRepo/search/query.html")
         self.assertEqual(len(response.context["res"]), 0)
         self.assertEqual(response.context["qry"], qry)
-
-    def test_pathStepToPosGroupType_inner_node(self):
-        """
-        Convert "0-all" to [0, "all"]
-        """
-        [pos, gtype, static] = pathStepToPosGroupType("0-all-True")
-        self.assertEqual(pos, 0)
-        self.assertEqual(gtype, "all")
-        self.assertTrue(not static)
-
-    def test_pathStepToPosGroupType_leaf_node(self):
-        """
-        Convert "0" to [0, None]
-        """
-        [pos, gtype, static] = pathStepToPosGroupType("0")
-        self.assertEqual(pos, 0)
-        self.assertEqual(gtype, None)
-        self.assertEqual(static, False)
-
-    def test_rootToFormatInfo_selected(self):
-        """
-        Convert "pgtemplate-PeakGroups-selected" to ["pgtemplate", "PeakGroups", True]
-        """
-        [format, name, sel] = rootToFormatInfo("pgtemplate-PeakGroups-selected")
-        self.assertEqual(format, "pgtemplate")
-        self.assertEqual(name, "PeakGroups")
-        self.assertEqual(sel, True)
-
-    def test_rootToFormatInfo_unselected(self):
-        """
-        Convert "pdtemplate-PeakData" to ["pdtemplate", "PeakData", False]
-        """
-        [format, name, sel] = rootToFormatInfo("pdtemplate-PeakData")
-        self.assertEqual(format, "pdtemplate")
-        self.assertEqual(name, "PeakData")
-        self.assertEqual(sel, False)
-
-    def test_search_advanced_tsv(self):
-        """
-        Download a simple advanced search and make sure the results are correct
-        """
-        [filledform, qry, dlform] = self.get_advanced_search_inputs()
-        response = self.client.post("/DataRepo/search_advanced_tsv/", dlform)
-
-        # Response content settings
-        self.assertEqual(response.get("Content-Type"), "application/text")
-        self.assertEqual(response.status_code, 200)
-        # Cannot use assertContains here for non-http response - it will complain about a missing status_code
-        contentdisp = response.get("Content-Disposition")
-        self.assertTrue("attachment" in contentdisp)
-        self.assertTrue("PeakGroups" in contentdisp)
-        self.assertTrue(".tsv" in contentdisp)
-
-    def test_getAllBrowseData(self):
-        """
-        Test that test_getAllBrowseData returns all data for the selected format.
-        """
-        basv_metadata = BaseAdvancedSearchView()
-        pf = "msrun__sample__animal__studies"
-        qs = PeakGroup.objects.all().prefetch_related(pf)
-        res, cnt = getAllBrowseData("pgtemplate", basv_metadata)
-        self.assertEqual(cnt, qs.count())
-
-    def get_basic_qry_inputs(self):
-        qs = PeakGroup.objects.all().prefetch_related("msrun__sample__animal__studies")
-        tval = str(qs[0].msrun.sample.animal.studies.all()[0].id)
-        empty_tree = {
-            "type": "group",
-            "val": "all",
-            "static": False,
-            "queryGroup": [
-                {
-                    "type": "query",
-                    "pos": "",
-                    "static": False,
-                    "ncmp": "",
-                    "fld": "",
-                    "val": "",
-                }
-            ],
-        }
-        qry = {
-            "selectedtemplate": "pgtemplate",
-            "searches": {
-                "pgtemplate": {
-                    "name": "PeakGroups",
-                    "tree": {
-                        "type": "group",
-                        "val": "all",
-                        "static": False,
-                        "queryGroup": [
-                            {
-                                "type": "query",
-                                "pos": "",
-                                "static": False,
-                                "ncmp": "iexact",
-                                "fld": "msrun__sample__animal__studies__name",
-                                "val": "obob_fasted",
-                            }
-                        ],
-                    },
-                },
-                "pdtemplate": {"name": "PeakData", "tree": empty_tree},
-                "fctemplate": {"name": "Fcirc", "tree": empty_tree},
-            },
-        }
-        return tval, qry
-
-    def test_createNewBasicQuery(self):
-        """
-        Test createNewBasicQuery creates a correct qry
-        """
-        tval, qry = self.get_basic_qry_inputs()
-        basv_metadata = BaseAdvancedSearchView()
-        mdl = "Study"
-        fld = "id"
-        cmp = "iexact"
-        val = tval
-        fmt = "pgtemplate"
-        newqry = createNewBasicQuery(basv_metadata, mdl, fld, cmp, val, fmt)
-        self.maxDiff = None
-        self.assertEqual(newqry, qry)
-
-    def test_searchFieldToDisplayField(self):
-        """
-        Test that searchFieldToDisplayField converts Study.id to Study.name
-        """
-        [tval, qry] = self.get_basic_qry_inputs()
-        qry["searches"]["pgtemplate"]["tree"]["queryGroup"][0][
-            "fld"
-        ] = "msrun__sample__animal__studies__id"
-        qry["searches"]["pgtemplate"]["tree"]["queryGroup"][0]["val"] = tval
-        basv_metadata = BaseAdvancedSearchView()
-        mdl = "Study"
-        fld = "id"
-        val = tval
-        dfld, dval = searchFieldToDisplayField(basv_metadata, mdl, fld, val, qry)
-        self.assertEqual(dfld, "name")
-        self.assertEqual(dval, "obob_fasted")
-
-    def test_getJoinedRecFieldValue(self):
-        """
-        Test that getJoinedRecFieldValue gets a value from a joined table
-        """
-        basv_metadata = BaseAdvancedSearchView()
-        fmt = "pgtemplate"
-        mdl = "Animal"
-        fld = "feeding_status"
-        pf = "msrun__sample__animal__studies"
-        recs = PeakGroup.objects.all().prefetch_related(pf)
-        val = getJoinedRecFieldValue(recs, basv_metadata, fmt, mdl, fld, fld, "Fasted")
-        self.assertEqual(val, "Fasted")
-
-    def test_constructAdvancedQuery(self):
-        """
-        Test that constructAdvancedQuery returns a correct Q expression
-        """
-        qry = self.get_advanced_qry()
-        q_exp = constructAdvancedQuery(qry)
-        expected_q = Q(msrun__sample__tissue__name__iexact="Brain")
-        self.assertEqual(q_exp, expected_q)
-
-    def test_performQuery(self):
-        """
-        Test that performQuery returns a correct queryset
-        """
-        qry = self.get_advanced_qry()
-        basv_metadata = BaseAdvancedSearchView()
-        pf = [
-            "msrun__sample__tissue",
-            "msrun__sample__animal__tracer_compound",
-            "msrun__sample__animal__studies",
-        ]
-        res, cnt = performQuery(qry, "pgtemplate", basv_metadata)
-        qs = PeakGroup.objects.filter(
-            msrun__sample__tissue__name__iexact="Brain"
-        ).prefetch_related(*pf)
-        self.assertEqual(cnt, qs.count())
-
-    def test_performQuery_distinct(self):
-        """
-        Test that performQuery returns no duplicate root table records when M:M tables queried with multiple matches.
-        """
-        qry = self.get_advanced_qry2()
-        basv_metadata = BaseAdvancedSearchView()
-        res, cnt = performQuery(qry, "pgtemplate", basv_metadata)
-        qs = (
-            PeakGroup.objects.filter(msrun__sample__name__iexact="BAT-xz971")
-            .filter(msrun__sample__animal__studies__name__iexact="obob_fasted")
-            .filter(compounds__synonyms__name__iexact="glucose")
-        )
-        # Ensure the test is working by ensuring the number of records without distinct is larger
-        self.assertTrue(cnt < qs.count())
-        self.assertEqual(cnt, 1)
-
-    def test_isQryObjValid(self):
-        """
-        Test that isQryObjValid correctly validates a qry object.
-        """
-        qry = self.get_advanced_qry()
-        basv_metadata = BaseAdvancedSearchView()
-        isvalid = isQryObjValid(qry, basv_metadata.getFormatNames().keys())
-        self.assertEqual(isvalid, True)
-        qry.pop("selectedtemplate")
-        isvalid = isQryObjValid(qry, basv_metadata.getFormatNames().keys())
-        self.assertEqual(isvalid, False)
-
-    def test_isValidQryObjPopulated(self):
-        """
-        Test that isValidQryObjPopulated correctly interprets the population of a subgroup.
-        """
-        qry = self.get_advanced_qry2()
-        isvalid = isValidQryObjPopulated(qry)
-        self.assertEqual(isvalid, True)
-        qry["searches"]["pgtemplate"]["tree"]["queryGroup"][1]["queryGroup"] = []
-        isvalid = isValidQryObjPopulated(qry)
-        self.assertEqual(isvalid, False)
 
     def test_cv_getSearchFieldChoices(self):
         """
