+++ DataRepo/urls.py
@@ -60,6 +60,14 @@ urlpatterns = [
         name="peakgroup_detail",
     ),
     path("peakdata/", views.PeakDataListView.as_view(), name="peakdata_list"),
-    path("task_status/<str:task_id>/", views.get_advsrch_download_progress, name='task_status'),
-    path("task_output/<str:task_id>/", views.get_advsrch_download_output, name='task_output'),
+    path(
+        "task_status/<str:task_id>/",
+        views.get_advsrch_download_progress,
+        name="task_status",
+    ),
+    path(
+        "task_output/<str:task_id>/",
+        views.get_advsrch_download_output,
+        name="task_output",
+    ),
 ]
